move DatabaseIO even deeper
